refactor: extract the secondary-label API into its own module
`dir_graph/mod.rs` was over its 2500-line ceiling, and this sprint's label
work had pushed it further out. Splitting rather than widening the allowance:
a tripped line ceiling means the file wants dividing.

Labels are the natural seam. They are the one piece of node state that does
not live in the storage backend at all — `NodeData` carries none, and
`DirGraph::secondary_label_index` is the canonical store — which is exactly
why the mutators are choke points rather than plain field access. Nothing
below them can observe a label change, so statement rollback and WAL capture
both have to be notified explicitly, and a write that bypassed them would be
silently unrollbackable and silently lost on crash recovery. That reasoning
now lives with the code it constrains.

Behaviour-neutral: the same `impl DirGraph` methods, moved. The label
crash-recovery tests (`test_secondary_labels_survive_crash`,
`test_labels_survive_checkpoint_then_crash`, both SIGKILL parameterisations)
pass unchanged, which is the point of running them against a move like this.
